lib/mb_mgr_burst_async: Cover full queue still processing corner case.
It is possible for the MB_MGR job queue to have all its jobs busy,
so `submit_burst_and_check()` will not be able to return number
of completions greater than 0. In that case, instead of marking
the queue as "empty", just return 0 and check if we ran into
this corner case during job flush.

Signed-off-by: Krzysztof Karas <krzysztof.karas@intel.com>
Change-Id: Ieba70b7ee6b20853584222d9f17687c15f27e22d
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/intel-ipsec-mb/+/19973
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Tested-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Ben Walker <ben@nvidia.com>